10/10 lkw vo삭제
