--- src/main/java/com/kosa/resq/domain/car/AccidentImgVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.car;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AccidentImgVO {
-    private String accident_code;
-    private String url;
-}
