--- src/main/java/com/kosa/resq/domain/car/CarDTO.java
@@ -1,4 +1,0 @@
-package com.kosa.resq.domain.car;
-
-public class CarDTO {
-}
