--- src/main/java/com/kosa/resq/domain/car/CarRezVO.java
@@ -1,19 +1,0 @@
-package com.kosa.resq.domain.car;
-
-import com.kosa.resq.domain.common.MemVO;
-
-import java.util.Date;
-
-public class CarRezVO {
-    private String car_rez_code;
-    private Date rez_at;
-    private String detail;
-    private String start_loc;
-    private String end_loc;
-    private float est_mileage;
-    private float rez_status;
-    private Date start_at;
-    private int rental_time;
-    private MemVO mem;
-    private CarVO car;
-}
