--- src/main/java/com/kosa/resq/domain/car/InsuranceComVO.java
@@ -1,14 +1,0 @@
-package com.kosa.resq.domain.car;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class InsuranceComVO {
-    private String ic_code;
-    private String ic_name;
-    private String ic_tel;
-}
