--- src/main/java/com/kosa/resq/domain/car/MaintComVO.java
@@ -1,14 +1,0 @@
-package com.kosa.resq.domain.car;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MaintComVO {
-    private String mc_code;
-    private String mc_name;
-    private String mc_tell;
-}
