--- src/main/java/com/kosa/resq/domain/common/AccountVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AccountVO {
-    private String account_code;
-    private String account_name;
-}
