--- src/main/java/com/kosa/resq/domain/common/AlertMsgVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AlertMsgVO {
-    private String msg_code;
-    private String contents;
-}
