--- src/main/java/com/kosa/resq/domain/common/PositionVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class PositionVO {
-    private String position_code;
-    private String position_name;
-}
