+++ src/main/java/com/kosa/resq/domain/dto/car/CarDTO.java
@@ -1,0 +1,4 @@
+package com.kosa.resq.domain.dto.car;
+
+public class CarDTO {
+}
