--- src/main/java/com/kosa/resq/domain/mr/BmGroupVO.java
@@ -1,14 +1,0 @@
-package com.kosa.resq.domain.mr;
-
-import com.kosa.resq.domain.common.MemVO;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class BmGroupVO {
-    private String bm_group_code;
-    private MemVO memVO;
-}
