--- src/main/java/com/kosa/resq/domain/mr/MrDTO.java
@@ -1,4 +1,0 @@
-package com.kosa.resq.domain.mr;
-
-public class MrDTO {
-}
