--- src/main/java/com/kosa/resq/domain/mr/MrImgVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.mr;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MrImgVO {
-    private MrVO mrVO;
-    private String img_url;
-}
