--- src/main/java/com/kosa/resq/domain/mr/MrKeywordVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.mr;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MrKeywordVO {
-    private MrVO mrVO;
-    private String keyword_name;
-}
