--- src/main/java/com/kosa/resq/domain/mr/MrOpDayVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.mr;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MrOpDayVO {
-    private MrVO mrVO;
-    private int day;
-}
