--- src/main/java/com/kosa/resq/domain/mr/MrTypeVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.mr;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MrTypeVO {
-    private String mr_type_code;
-    private String type_name;
-}
