--- src/main/java/com/kosa/resq/domain/mr/OutsiderVO.java
@@ -1,14 +1,0 @@
-package com.kosa.resq.domain.mr;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class OutsiderVO {
-    private String outsider_code;
-    private String name;
-    private String company_name;
-}
