--- src/main/java/com/kosa/resq/domain/mr/SuppliesImgVO.java
@@ -1,13 +1,0 @@
-package com.kosa.resq.domain.mr;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SuppliesImgVO {
-    private SuppliesVO suppliesVO;
-    private String img_url;
-}
