+++ src/main/java/com/kosa/resq/domain/vo/car/CarVO.java
@@ -1,0 +1,4 @@
+package com.kosa.resq.domain.vo.car;
+
+public class CarVO {
+}
