+++ src/main/java/com/kosa/resq/domain/vo/common/MemVO.java
@@ -1,0 +1,4 @@
+package com.kosa.resq.domain.vo.common;
+
+public class MemVO {
+}
