+++ src/main/java/com/kosa/resq/domain/vo/mr/MrVO.java
@@ -1,0 +1,4 @@
+package com.kosa.resq.domain.vo.mr;
+
+public class MrVO {
+}
