--- src/main/java/com/kosa/resq/service/car/CarUserService.java
@@ -1,5 +1,0 @@
-package com.kosa.resq.service.car;
-
-
-public interface CarUserService {
-}
