+++ src/main/resources/mapper/car/CarUserMapper.xml
@@ -3,46 +3,9 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.kosa.resq.mapper.car.CarUserMapper">
-    <resultMap id="CarRezMap" type="domain.car.CarRezVO">
-        <id property="car_rez_code" column="car_rez_code"/>
-        <result property="car_rez_code" column="car_rez_code"/>
-        <result property="rez_at" column="rez_at"/>
-        <result property="detail" column="detail"/>
-        <result property="start_loc" column="end_loc"/>
-        <result property="est_mileage" column="est_mileage"/>
-        <result property="rez_status" column="rez_status"/>
-        <result property="rental_time" column="rental_time"/>
-        <collection property="mem" resultMap="MemMap"/>
-        <collection property="car" resultMap="CarMap"/>
-    </resultMap>
-    <resultMap id="CarMap" type="domain.car.CarVO">
-        <id property="" column=""/>
-        <result property="" column=""/>
-    </resultMap>
-    <resultMap id="MemMap" type="domain.common.MemVO">
-        <id property="mem_code" column="mem_code"/>
-        <result property="mem_code" column="mem_code"/>
-        <result property="name" column="name"/>
-        <result property="email" column="password"/>
-        <result property="role" column="role"/>
-        <result property="hired_at" column="hired_at"/>
-        <result property="created_at" column="created_at"/>
-        <result property="address" column="address"/>
-        <result property="al_date_cnt" column="al_date_cnt"/>
-        <result property="tel" column="tel"/>
-        <result property="ssn" column="ssn"/>
-        <result property="is_worked" column="is_worked"/>
-        <result property="profile_img_url" column="profile_img_url"/>
-        <collection property="dept" resultMap="DeptMap"/>
-        <collection property="position" resultMap="PositionMap"/>
-    </resultMap>
-    <resultMap id="DeptMap" type="domain.common.DeptVO">
-        
-    </resultMap>
-    <resultMap id="PositionMap" type="domain.common.PositionVO">
 
-    </resultMap>
-    <insert id="CarRezSave" parameterMap="CarRezMap">
+    <select id="carRezSeq" resultType="Integer">
+        select car_rez_seq.nextval from dual
+    </select>
 
-    </insert>
 </mapper>
